old records cleaner class
